requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
abseil-cpp=20230802.0=h313beb8_2
absl-py=1.4.0=py39hca03da5_0
aiohttp=3.9.0=py39h80987f9_0
aiosignal=1.2.0=pyhd3eb1b0_0
antlr4-python3-runtime=4.9.3=pypi_0
arrow-cpp=11.0.0=hc7aafb3_2
async-timeout=4.0.3=py39hca03da5_0
attrs=23.1.0=py39hca03da5_0
aws-c-common=0.6.8=h80987f9_1
aws-c-event-stream=0.1.6=h313beb8_6
aws-checksums=0.1.11=h80987f9_2
aws-sdk-cpp=1.8.185=ha71a6ea_1
blas=1.0=openblas
blinker=1.6.2=py39hca03da5_0
boost-cpp=1.82.0=h48ca7d4_2
bottleneck=1.3.5=py39heec5a64_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
brotli-python=1.0.9=py39hc377ac9_7
bzip2=1.0.8=h620ffc9_4
c-ares=1.19.1=h80987f9_0
ca-certificates=2023.12.12=hca03da5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.11.17=py39hca03da5_0
cffi=1.16.0=py39h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39hca03da5_0
contourpy=1.2.0=py39h48ca7d4_0
crc32c=2.3.post0=pypi_0
cryptography=41.0.7=py39hd4332d6_0
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.16.1=pyhd8ed1ab_0
dill=0.3.6=py39hca03da5_0
einops=0.7.0=pypi_0
ffmpeg=4.2.2=h04105a8_0
filelock=3.13.1=py39hca03da5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
frozenlist=1.4.0=py39h80987f9_0
fsspec=2023.10.0=py39hca03da5_0
gettext=0.21.0=h13f89a0_1
gflags=2.2.2=hc377ac9_0
giflib=5.2.1=h80987f9_3
glog=0.5.0=hc377ac9_0
gmp=6.2.1=hc377ac9_3
gmpy2=2.1.2=py39h8c48613_0
gnutls=3.6.15=h887c41c_0
google-auth=2.22.0=py39hca03da5_0
google-auth-oauthlib=0.5.2=py39hca03da5_0
grpc-cpp=1.48.2=hc60591f_4
grpcio=1.48.2=py39hc60591f_4
gtest=1.14.0=h48ca7d4_0
huggingface_hub=0.20.2=pyhd8ed1ab_0
hydra-core=1.3.2=pypi_0
icu=73.1=h313beb8_0
idna=3.4=py39hca03da5_0
importlib-metadata=7.0.1=py39hca03da5_0
importlib_resources=6.1.1=py39hca03da5_1
jinja2=3.1.2=py39hca03da5_0
joblib=1.2.0=py39hca03da5_0
jpeg=9e=h80987f9_1
kiwisolver=1.4.4=py39h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libboost=1.82.0=h0bc93f9_2
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcurl=8.5.0=h3e2b118_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libevent=2.1.12=h02f6b3c_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.4=h80987f9_0
libjpeg-turbo=2.0.0=h1a28f6b_0
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.39=h80987f9_0
libprotobuf=3.20.3=h514c7bf_0
libssh2=1.10.0=h02f6b3c_2
libtasn1=4.19.0=h80987f9_0
libthrift=0.15.0=h73c2103_2
libtiff=4.5.1=h313beb8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.4=h0dcf63f_1
lightning=2.1.3=pyhd8ed1ab_1
lightning-utilities=0.10.0=pyhd8ed1ab_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
markdown=3.4.1=py39hca03da5_0
markupsafe=2.1.3=py39h80987f9_0
matplotlib=3.8.0=py39hca03da5_0
matplotlib-base=3.8.0=py39h46d7db6_0
mpc=1.1.0=h8c48613_1
mpfr=4.0.2=h695f6f0_1
mpmath=1.3.0=py39hca03da5_0
multidict=6.0.4=py39h80987f9_0
multiprocess=0.70.14=py39hca03da5_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
nettle=3.7.3=h84b5d62_1
networkx=3.1=py39hca03da5_0
numexpr=2.8.7=py39hecc3335_0
numpy=1.26.3=py39h3b2db8e_0
numpy-base=1.26.3=py39ha9811e2_0
oauthlib=3.2.2=py39hca03da5_0
omegaconf=2.3.0=pypi_0
openh264=1.8.0=h98b2900_0
openjpeg=2.3.0=h7a6adac_2
openssl=3.0.12=h1a28f6b_0
orc=1.7.4=hdca1487_1
packaging=23.1=py39hca03da5_0
pandas=2.1.4=py39h46d7db6_0
pillow=10.0.1=py39h3b245a6_0
pip=23.3.1=py39hca03da5_0
protobuf=3.20.3=py39h313beb8_0
pyarrow=11.0.0=py39hbfed03b_1
pyarrow-hotfix=0.6=pyhd8ed1ab_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py39hca03da5_0
pyopenssl=23.2.0=py39hca03da5_0
pyparsing=3.0.9=py39hca03da5_0
pysocks=1.7.1=py39hca03da5_0
python=3.9.18=hb885b13_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h1a28f6b_1
pytorch=2.1.2=py3.9_0
pytorch-lightning=2.1.3=pyhd8ed1ab_0
pytz=2023.3.post1=py39hca03da5_0
pyyaml=6.0.1=py39h80987f9_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
regex=2023.10.3=py39h80987f9_0
requests=2.31.0=py39hca03da5_0
requests-oauthlib=1.3.0=py_0
responses=0.13.3=pyhd3eb1b0_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.0=py39h482802a_0
scikit-learn=1.3.0=py39h46d7db6_1
scipy=1.11.4=py39h20cbe94_0
setuptools=68.2.2=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h313beb8_1
sqlite=3.41.2=h80987f9_0
sympy=1.12=py39hca03da5_0
tensorboard=2.12.1=py39hca03da5_0
tensorboard-data-server=0.7.0=py39ha6e5c4f_0
tensorboard-plugin-wit=1.8.1=py39hca03da5_0
tfrecord=1.14.4=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.13.2=py39h3dd52b7_1
torchaudio=2.1.2=py39_cpu
torchmetrics=1.2.1=pyhd8ed1ab_0
torchvision=0.16.2=py39_cpu
tornado=6.3.3=py39h80987f9_0
tqdm=4.65.0=py39h86d0a89_0
transformers=4.32.1=py39hca03da5_0
typing-extensions=4.7.1=py39hca03da5_0
typing_extensions=4.7.1=py39hca03da5_0
tzdata=2023d=h04d1e81_0
urllib3=1.26.18=py39hca03da5_0
utf8proc=2.6.1=h1a28f6b_0
werkzeug=2.2.3=py39hca03da5_0
wheel=0.41.2=py39hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xxhash=0.8.0=h1a28f6b_3
xz=5.4.5=h80987f9_0
yaml=0.2.5=h1a28f6b_0
yarl=1.9.3=py39h80987f9_0
zipp=3.17.0=py39hca03da5_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
